Fix usage of getUpdates offset
Without this patch slimbot increments getUpdates offset every time getUpdates
is called. So the offset value is usually greater then the received updates id.
But the doc says that offset is:

"Identifier of the first update to be returned. Must be greater by one than
the highest among the identifiers of previously received updates. By default,
updates starting with the earliest unconfirmed update are returned. An update
is considered confirmed as soon as getUpdates is called with an offset higher
than its update_id."

The current behaviour is not in conformance to the doc and seems to result in
losing updates from time to time, which I observe in my bot. Here are my
thoughts on how it happens:

1. At first offset is 0. A library user calls startPolling(). startPolling()
   increases offset by 1, starts a getUpdates request and, when it finishes,
   starts a timer to call itself again in 100ms. So if there are no updates
   offset keeps increasing like this: 1, 2, 3, ...
2. Then there are some updates. Offset is set to the update_id field value of
   the last processed update. BTW update_id field values are increasing by 1
   with each update according to the spec (unless there is no activity for a
   week). Let's assume the offset is now 1000.
3. Then startPolling() is called again in loop. Thus offset keeps increasing
   like this: 1001, 1002, 1003, ...
4. Then there is one more update. It's update_id should be 1001 according to
   the spec. If the update comes during the time the getUpdates request is
   running, then for some unclear reason it is delivered. Probably telegram
   devs didn't expect things like "skip first n updates" to be useful :)
   But in a rare case when getUpdates is not running (during 100ms timeout,
   when sending the next request over network, node.js event loop is busy
   with other stuff) the update is stored as unconfirmed, then the next
   getUpdates request with a greater offset just discards the update.

So we have at least 1% probability of losing updates.